assistant-toolkit: agent-chat implementation module behind namespace facades; bust react-ui build cache
The import-as-namespace pragma forbids named re-exports from Chat.ts, so the
co-located pair moves to a pragma-free internal module with both namespaces as
facades. The react-ui comment touch invalidates a corrupt remote-cache dist
(chunk-ErrorStack.mjs missing in CI; local build and 163 storybook tests green).
